Rename Tests to Auto run by maven
