+++ src/test/java/de/viadee/bpm/camunda/externaltask/retry/aspect/behaviour/DefaultBehaviourForExceptionTypesTest.java
@@ -42,7 +42,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
 
 
-public class DefaultBehaviourForExceptionTypesTestCamunda extends CamundaBaseTest {
+public class DefaultBehaviourForExceptionTypesTest extends CamundaBaseTest {
 
 
     @Test
