+++ src/test/java/de/viadee/bpm/camunda/externaltask/retry/aspect/behaviour/InvalidRetryTimeCycleValuesTest.java
@@ -37,15 +37,29 @@ import org.springframework.test.context.TestPropertySource;
 
 import static org.mockito.Mockito.when;
 
-@TestPropertySource(properties = "de.viadee.bpm.camunda.external-task.retry-config.default-behavior=R3/PT1D")
-public class InvalidDefaultBehaviourTestCamunda extends CamundaBaseTest {
+
+@TestPropertySource(properties = "de.viadee.bpm.camunda.external-task.retry-config.default-behavior=R3/PT37M")
+public class InvalidRetryTimeCycleValuesTest extends CamundaBaseTest {
 
 
     @Test
-    public void runtimeException() {
+    public void invalidRetryTimeCycle() {
+        // PT3D not valid -> 'default-retry-time-cycle' should be used
+        this.invalidTimeCycleDetectedByRegularExpression("R3/PT3D");
+    }
+
+
+    @Test
+    public void invalidRetryTimeCycleList() {
+        // PT3D not valid -> 'default-retry-time-cycle' should be used
+        this.invalidTimeCycleDetectedByRegularExpression("PT10M,PT3D,PT10M");
+    }
+
+
+    public void invalidTimeCycleDetectedByRegularExpression(final String retryTimeCycle) {
         // prepare
-        when(this.externalTask.getRetries()).thenReturn(null);
-        when(this.externalTask.getExtensionProperty(this.properties.getIdentifier())).thenReturn("invld!");
+        when(this.externalTask.getRetries()).thenReturn(3);
+        when(this.externalTask.getExtensionProperty(this.properties.getIdentifier())).thenReturn(retryTimeCycle);
 
         // test
         this.camundaExternalTaskRetryAspect.handleErrorAfterThrown(this.joinPoint, new RuntimeException(), this.externalTask, this.externalTaskService);
@@ -55,9 +69,8 @@ public class InvalidDefaultBehaviourTestCamunda extends CamundaBaseTest {
         this.verifyHandleFailure();
 
         // assert
-        this.assertRemainingRetries(3);
-        this.assertNextRetryInterval(5 * MINUTES_TO_MILLIS);
+        this.assertRemainingRetries(2);
+        this.assertNextRetryInterval(37 * MINUTES_TO_MILLIS);
     }
-
 
 }
